Move local_vote_history in to its own file.
