+++ nano/core_test/voting.cpp
@@ -1,5 +1,6 @@
 #include <nano/lib/blocks.hpp>
 #include <nano/node/common.hpp>
+#include <nano/node/local_vote_history.hpp>
 #include <nano/node/voting.hpp>
 #include <nano/secure/ledger.hpp>
 #include <nano/test_common/system.hpp>
